PLUS-trunk:  * REQDEL (See #179)    * Update error message when failing to delete a dir because it is empty, clean up some of the logic to make it easier to read.    * While I'm here anyway, fix it so that it credits request count for user who owns the request, instead of the user who initiated REQDEL command (in the case of an admin)
git-svn-id: https://drftpd.svn.sourceforge.net/svnroot/drftpd/plus/2.0/trunk@1821 c83153ee-f3e8-0310-b165-dc3a72e09794

+++ src/net/drmods/plugins/irc/Request.java
@@ -36,6 +36,7 @@ import org.drftpd.dynamicdata.Key;
 import org.drftpd.plugins.SiteBot;
 import org.drftpd.remotefile.LinkedRemoteFileInterface;
 import org.drftpd.sitebot.IRCCommand;
+import org.drftpd.usermanager.NoSuchUserException;
 import org.drftpd.usermanager.User;
 import org.drftpd.usermanager.UserFileException;
 import org.tanesha.replacer.ReplacerEnvironment;
@@ -315,27 +316,38 @@ public class Request extends IRCCommand {
             
             for (Iterator iter = dirs.iterator(); iter.hasNext();) {
                 LinkedRemoteFileInterface file = (LinkedRemoteFileInterface) iter.next();
-                if (file.isDirectory()) {
-                    if (file.getName().endsWith(dirName)) {
-                        nodir = false;
-                        if (file.dirSize() > 0 && (file.getUsername().equals(user.getName()) || user.isAdmin())) {
-                            file.delete();
-                            deldir = true;
-                            ArrayList<String> forceToChannels = new ArrayList<String>();
-                            forceToChannels.add(msgc.getDest());
-                            getGlobalContext().dispatchFtpEvent(new DirectorySiteBotEvent(
-                            		user, "REQDEL", file, forceToChannels));
-                            // decrement request counter
-                            int reqsMade = user.getKeyedMap().getObjectInt(Request.REQUESTS);
-                            if (reqsMade > 0)
-                            	user.getKeyedMap().setObject(Request.REQUESTS, --reqsMade);
-                            break;
-                        } else {
-                            out.add(ReplacerUtils.jprintf("reqdel.notowner", env, Request.class));
-                            break;
-                        }
-                    } else nodir = true;
-                }
+                if (file.isDirectory() && file.getName().endsWith(dirName)) {
+                    nodir = false;
+                    if (file.dirSize() > 0) {
+                    	if (user.isAdmin() || file.getUsername().equals(user.getName())) {
+	                    	User requestowner = null;
+								try {
+									requestowner = getGlobalContext().getUserManager().getUserByName(file.getUsername());
+								} catch (NoSuchUserException e) {
+								} catch (UserFileException e) {
+								}
+	                        file.delete();
+	                        deldir = true;
+	                        ArrayList<String> forceToChannels = new ArrayList<String>();
+	                        forceToChannels.add(msgc.getDest());
+	                        getGlobalContext().dispatchFtpEvent(new DirectorySiteBotEvent(
+	                        		user, "REQDEL", file, forceToChannels));
+	                        // decrement request counter
+	                        if (requestowner != null) {
+	                        	int reqsMade = requestowner.getKeyedMap().getObjectInt(Request.REQUESTS);
+	                        	if (reqsMade > 0)
+	                        		requestowner.getKeyedMap().setObject(Request.REQUESTS, --reqsMade);
+	                        }
+	                        break;
+	                    } else {
+	                        out.add(ReplacerUtils.jprintf("reqdel.notowner", env, Request.class));
+	                        break;
+	                    }
+                    } else {
+                        out.add(ReplacerUtils.jprintf("reqdel.notempty", env, Request.class));
+                        break;
+	                }
+                } else nodir = true;
             }
             
             if (nodir && !deldir)
